requirements.txt
anyio==3.6.2
certifi==2023.5.7
flake8==6.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
